App: Finished refining the Task list and Task detail UI.
