--- app/app/src/main/java/com/example/msc_onlab/ui/feature/tasks/TasksBottomSheet.kt
@@ -1,113 +1,0 @@
-package com.example.msc_onlab.ui.feature.tasks
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Delete
-import androidx.compose.material.icons.rounded.Edit
-import androidx.compose.material.icons.rounded.FileOpen
-import androidx.compose.material.icons.rounded.Share
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.SheetState
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun TasksBottomSheet(
-    taskTitle: String,
-    onDismissRequest: () -> Unit,
-    sheetState: SheetState,
-    onEdit: () -> Unit,
-    onDelete: () -> Unit,
-    onDetails: () -> Unit,
-){
-    ModalBottomSheet(
-        onDismissRequest = { onDismissRequest() },
-        sheetState = sheetState
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(15.dp)
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(50.dp)
-                    .clickable { },
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = taskTitle,
-                    fontSize = 18.sp,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-            HorizontalDivider()
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(50.dp)
-                    .clickable { onEdit() },
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.Start
-            ) {
-                Icon(imageVector = Icons.Rounded.Edit, contentDescription = "Edit")
-                Spacer(modifier = Modifier.padding(horizontal = 10.dp))
-                Text(
-                    text = "Edit task",
-                    fontSize = 18.sp
-                )
-            }
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(50.dp)
-                    .clickable { onDetails() },
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.Start
-            ) {
-                Icon(imageVector = Icons.Rounded.FileOpen, contentDescription = "Details")
-                Spacer(modifier = Modifier.padding(horizontal = 10.dp))
-                Text(
-                    text = "Show details",
-                    fontSize = 18.sp
-                )
-            }
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(50.dp)
-                    .clickable { onDelete() },
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.Start
-            ) {
-                Icon(imageVector = Icons.Rounded.Delete, contentDescription = "Delete")
-                Spacer(modifier = Modifier.padding(horizontal = 10.dp))
-                Text(
-                    text = "Delete task",
-                    fontSize = 18.sp
-                )
-            }
-        }
-    }
-}
